refactor handlers separados por archivo y exportar sendEmail
